+++ R/add_dummy_from_dummies.R
@@ -5,6 +5,8 @@
 #' @param ... Dummy columns from df, must be in set c(0,1).
 #' @param na_to_zero Boolean. Mutate NA values to zero.
 #'
+#' @importFrom rlang `:=`
+#'
 #' @return An update data frame with a new dummy column.
 #' @export
 add_dummy_from_dummies <- function(df,
@@ -26,7 +28,7 @@ add_dummy_from_dummies <- function(df,
       "{dummy_name}" := dplyr::case_when(
         rowSums(dplyr::across(dplyr::all_of(quoted_cols)), na.rm = TRUE) >= 1 ~ 1,
         rowSums(dplyr::across(dplyr::all_of(quoted_cols)), na.rm = TRUE) < 1 ~ 0,
-        TRUE ~ NA_real_
+        .default = NA_real_
       )
     )
   } else {
@@ -35,7 +37,7 @@ add_dummy_from_dummies <- function(df,
       "{dummy_name}" := dplyr::case_when(
         rowSums(dplyr::across(dplyr::all_of(quoted_cols)), na.rm = FALSE) >= 1 ~ 1,
         rowSums(dplyr::across(dplyr::all_of(quoted_cols)), na.rm = FALSE) < 1 ~ 0,
-        TRUE ~ NA_real_
+        .default = NA_real_
       )
     )
   }