+++ NEWS.md
@@ -1,3 +1,9 @@
+# humind 0.0.2
+
+* Added: HWISE functions. (#16)
+* Updated: to dplyr v1.1.0 with the `dplyr::case_when()` revamp. (#)
+* Modified: dependency know goes to `impactR.utils` and not `impactR`. (#)
+
 # humind 0.0.1
 
 * A few patches to github-action and pages